skystrike_fullstack_final_release/backend/requirements.txt
fastapi
pydantic
pyjwt
httpx
numpy
python-dotenv
passlib
